delete original behavior change
Changes behavior when original file is an mp4

+++ postConversion.py
@@ -29,7 +29,7 @@ if len(sys.argv) > 4:
     tagmp4.setHD(convert.width, convert.height)
     tagmp4.writeTags(path)
     if settings.output_dir is not None:
-        if extension in valid_output_extensions:  # If the file is already in a valid format, this will duplicate the file in the output directory since no original would be left behind
+        if extension in valid_output_extensions and settings.delete is False:  # If the file is already in a valid format, this will duplicate the file in the output directory since no original would be left behind
             output = os.path.join(settings.output_dir, os.path.split(path)[1])
             try:
                 shutil.copy(path, output)
